requirements.txt
h5py==2.10.0
matplotlib==2.2.5
numpy==1.16.6
scipy==1.2.3
SimpleITK==1.2.4
tables==3.5.2
tensorflow-gpu==1.14.0
pyaml
scikit-image